+++ src/execution/__tests__/defer-test.ts
@@ -18,7 +18,7 @@ import {
 import { GraphQLID, GraphQLString } from '../../type/scalars.js';
 import { GraphQLSchema } from '../../type/schema.js';
 
-import { execute, experimentalExecuteIncrementally } from '../execute.js';
+import { execute, experimentalExecuteIncrementally } from '../Executor.js';
 import type {
   InitialIncrementalExecutionResult,
   SubsequentIncrementalExecutionResult,
