+++ src/execution/__tests__/nonnull-test.ts
@@ -14,7 +14,7 @@ import { GraphQLSchema } from '../../type/schema.js';
 
 import { buildSchema } from '../../utilities/buildASTSchema.js';
 
-import { execute, executeSync } from '../execute.js';
+import { execute, executeSync } from '../Executor.js';
 import type { ExecutionResult } from '../types.js';
 
 const syncError = new Error('sync');
