+++ src/execution/__tests__/subscribe-test.ts
@@ -20,12 +20,12 @@ import {
 } from '../../type/scalars.js';
 import { GraphQLSchema } from '../../type/schema.js';
 
-import type { ExecutionArgs } from '../execute.js';
+import type { ExecutionArgs } from '../Executor.js';
 import {
   createSourceEventStream,
   executeSubscriptionEvent,
   subscribe,
-} from '../execute.js';
+} from '../Executor.js';
 import type { ExecutionResult } from '../types.js';
 
 import { SimplePubSub } from './simplePubSub.js';
